All files added
This is a "naive" version of Inception-V4 inference. I call it naive because Iit doesn't have support for minibatch or any optimization. It is intended to be converted to HLS sythesizable code. The code was finished long time ago, though I was procrastinating on pushing it to git. To save space, I removed all data files, the checkpoint file used can be acquired from http://download.tensorflow.org/models/inception_v4_2016_09_09.tar.gz